progress: change reduction type to uint64_t

+++ src/common/mfu_flist.h
@@ -493,8 +493,8 @@ typedef struct {
     time_t current;
     int timeout;
     int keep_going;
-    int values[2];      /* local array */
-    int global_vals[2]; /* global array across all ranks */
+    uint64_t values[2];      /* local array */
+    uint64_t global_vals[2]; /* global array across all ranks */
     int count;
 } mfu_progress_msgs_t;
 
@@ -504,12 +504,12 @@ mfu_progress_msgs_t* mfu_progress_start(int secs, MPI_Comm dupcomm);
 
 /* update progress across all processes in work loop */
 void mfu_progress_update(mfu_progress_msgs_t* msgs,
-                         int file_count);
+                         uint64_t file_count);
 
 /* continue broadcasting progress until all processes have completed,
  * and free structure allocated in start */
 void mfu_progress_complete(mfu_progress_msgs_t** pmsgs,
-                           int file_count);
+                           uint64_t file_count);
 
 /* given a file list and a chunk size, split files at chunk boundaries and evenly
  * spread chunks to processes, returns a linked list of file sections each process
